requirements.txt
streamlit
requests
fastapi
uvicorn
google-generativeai
PyPDF2
python-multipart